requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
dill==0.3.6
environs==9.5.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
magic-filter==1.0.9
marshmallow==3.19.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
platformdirs==3.5.1
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.9
pylint==2.17.4
python-dotenv==1.0.0
python-dotenv-vault==0.6.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.15
tomlkit==0.11.8
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.6.3
urllib3==2.0.2
wrapt==1.15.0
yarl==1.9.2
